Rename `Bordered*` components to `Border*` (#101)
* Rename component to `BorderdCircleButton`

* Rename component to `BorderRoundButton`

* Rename component to `BorderTitle`

* Rename test page of `border-title`

* Cleanup codes

* Fix `className` gets set to `undefined`
