+++ src/components/buttons/BorderCircleButton/styles.module.scss
@@ -14,7 +14,6 @@
 
     width: calc(100% - 0.375em);
     aspect-ratio: 1 / 1;
-    content: "";
     background-color: #fff;
   }
