+++ src/components/headers/BorderTitle/styles.module.scss
